# internal/api/router.go
package api

import (
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"net/http"
	"strconv"
	"time"

	"eats-backend/internal/config"
	"eats-backend/internal/models"

	"github.com/rs/cors"
	"go.uber.org/zap"
)

var (
	errInvalidPaginationParameter = errors.New("invalid pagination parameter")
	errEmptyID                    = errors.New("empty id")
	errEmptyName                  = errors.New("empty name")
)

type FileSaver interface {
	SaveFile(w http.ResponseWriter, r *http.Request)
}

type ProductsService interface {
	GetProductsList(ctx context.Context, page, pageSize int, category string) (models.ProductsList, error)
	GetProductByID(ctx context.Context, id string) (models.Product, error)
	GetCategories() []models.Category
	AddFavourite(ctx context.Context, id string) error
	RemoveFavourite(ctx context.Context, id string) error
}

type TokenService interface {
	GenerateToken(ctx context.Context, username string, isTeacher bool) (string, error)
}

type Router struct {
	*http.Server
	router *http.ServeMux

	productsService ProductsService
	tokenService    TokenService
	fileSaver       FileSaver

	logger *zap.SugaredLogger
}

func NewRouter(
	cfg config.ServerOpts,
	productsService ProductsService,
	tokenService TokenService,
	fileSaver FileSaver,
	authMiddleware func(next http.HandlerFunc) http.HandlerFunc,
	logger *zap.SugaredLogger,
) *Router {
	innerRouter := http.NewServeMux()

	appRouter := &Router{
		Server: &http.Server{
			Handler:      cors.AllowAll().Handler(innerRouter),
			ReadTimeout:  time.Duration(cfg.ReadTimeout) * time.Second,
			WriteTimeout: time.Duration(cfg.WriteTimeout) * time.Second,
			IdleTimeout:  time.Duration(cfg.IdleTimeout) * time.Second,
		},
		router:          innerRouter,
		productsService: productsService,
		tokenService:    tokenService,
		logger:          logger,
		fileSaver:       fileSaver,
	}

	innerRouter.HandleFunc("GET /users/me", authMiddleware(nil))
	innerRouter.HandleFunc("PUT /users/me", authMiddleware(nil))
	innerRouter.HandleFunc("DELETE /users/me", authMiddleware(nil))

	innerRouter.HandleFunc("POST /logout", authMiddleware(nil))

	innerRouter.HandleFunc("GET /products", authMiddleware(appRouter.getProductsList))
	innerRouter.HandleFunc("GET /products/{id}", authMiddleware(appRouter.getProductByID))

	innerRouter.HandleFunc("POST /products/{id}/favourite", authMiddleware(appRouter.addFavourite))
	innerRouter.HandleFunc("DELETE /products/{id}/favourite", authMiddleware(appRouter.deleteFavourite))

	innerRouter.HandleFunc("POST /products/{id}/reviews", authMiddleware(nil))

	innerRouter.HandleFunc("GET /categories", authMiddleware(appRouter.getCategories))

	innerRouter.HandleFunc("GET /cart", authMiddleware(nil))
	innerRouter.HandleFunc("POST /cart/items", authMiddleware(nil))
	innerRouter.HandleFunc("DELETE /cart/items/{id}", authMiddleware(nil))

	innerRouter.HandleFunc("GET /orders", authMiddleware(nil))
	innerRouter.HandleFunc("POST /orders", authMiddleware(nil))

	innerRouter.HandleFunc("GET /addresses", authMiddleware(nil))
	innerRouter.HandleFunc("POST /addresses", authMiddleware(nil))
	innerRouter.HandleFunc("PUT /addresses/{id}", authMiddleware(nil))

	innerRouter.HandleFunc("POST /createToken", authMiddleware(appRouter.createToken))
	innerRouter.HandleFunc("POST /createTeacherToken", authMiddleware(appRouter.createTeacherToken))

	uploadsDir := http.Dir("data/uploads")
	innerRouter.Handle("GET /uploads/", http.StripPrefix("/uploads/", http.FileServer(uploadsDir)))
	innerRouter.HandleFunc("POST /uploads", authMiddleware(appRouter.fileSaver.SaveFile))

	innerRouter.HandleFunc("GET /", func(writer http.ResponseWriter, request *http.Request) {
		http.ServeFile(writer, request, "redoc-static.html")
	})

	return appRouter
}

func (r *Router) sendResponse(response http.ResponseWriter, request *http.Request, code int, buf []byte) {
	response.Header().Set("Content-Type", "application/json")
	response.WriteHeader(code)
	_, err := response.Write(buf)
	if err != nil {
		r.logger.With(
			"module", "api",
			"request_url", request.Method+": "+request.URL.Path,
		).Errorf("Error sending error response: %v", err)
	}
}

func (r *Router) sendErrorResponse(response http.ResponseWriter, request *http.Request, err error) {
	switch {
	case errors.Is(err, models.ErrBadRequest):
		response.WriteHeader(http.StatusBadRequest)
		r.logger.With(
			"module", "api",
			"request_url", request.Method+": "+request.URL.Path,
		).Warn(err)
		r.writeError(response, request, err)

		return
	case errors.Is(err, models.ErrNotFound):
		response.WriteHeader(http.StatusNotFound)
		r.logger.With(
			"module", "api",
			"request_url", request.Method+": "+request.URL.Path,
		).Warn(err)

		r.writeError(response, request, err)

		return
	case errors.Is(err, models.ErrForbidden):
		response.WriteHeader(http.StatusForbidden)
		r.logger.With(
			"module", "api",
			"request_url", request.Method+": "+request.URL.Path,
		).Warn(err)

		r.writeError(response, request, err)

		return
	case errors.Is(err, models.ErrUnauthorized):
		response.WriteHeader(http.StatusUnauthorized)
		r.logger.With(
			"module", "api",
			"request_url", request.Method+": "+request.URL.Path,
		).Warn(err)

		r.writeError(response, request, err)

		return
	}

	response.WriteHeader(http.StatusInternalServerError)
	r.logger.With(
		"module", "api",
		"request_url", request.Method+": "+request.URL.Path,
	).Error(err)

	r.writeError(response, request, err)
}

func (r *Router) writeError(response http.ResponseWriter, request *http.Request, err error) {
	_, err = response.Write([]byte(err.Error()))
	if err != nil {
		r.logger.With(
			"module", "api",
			"request_url", request.Method+": "+request.URL.Path,
		).Errorf("Error sending error response: %v", err)
	}
}

func (r *Router) getProductsList(writer http.ResponseWriter, request *http.Request) {
	page, err := getPaginationParameter(request, "page", 1)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, err))

		return
	}

	pageSize, err := getPaginationParameter(request, "pageSize", models.DefaultPageSize)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, err))
	}

	category := request.URL.Query().Get("category")

	result, err := r.productsService.GetProductsList(request.Context(), page, pageSize, category)
	if err != nil {
		r.sendErrorResponse(writer, request, err)
	}

	buf, err := json.Marshal(result)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrInternalServer, err))

		return
	}

	r.sendResponse(writer, request, http.StatusOK, buf)
}

func (r *Router) getProductByID(writer http.ResponseWriter, request *http.Request) {
	id := request.PathValue("id")
	if id == "" {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, errEmptyID))

		return
	}

	product, err := r.productsService.GetProductByID(request.Context(), id)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("GetProductByID: %w", err))

		return
	}

	buf, err := json.Marshal(product)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrInternalServer, err))

		return
	}

	r.sendResponse(writer, request, http.StatusOK, buf)
}

func (r *Router) addFavourite(writer http.ResponseWriter, request *http.Request) {
	id := request.PathValue("id")
	if id == "" {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, errEmptyID))

		return
	}

	err := r.productsService.AddFavourite(request.Context(), id)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("AddFavourite: %w", err))

		return
	}

	writer.WriteHeader(http.StatusOK)
}

func (r *Router) deleteFavourite(writer http.ResponseWriter, request *http.Request) {
	id := request.PathValue("id")
	if id == "" {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, errEmptyID))

		return
	}

	err := r.productsService.RemoveFavourite(request.Context(), id)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("RemoveFavourite: %w", err))

		return
	}

	writer.WriteHeader(http.StatusOK)
}

func (r *Router) getCategories(writer http.ResponseWriter, request *http.Request) {
	result := r.productsService.GetCategories()

	buf, err := json.Marshal(result)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrInternalServer, err))

		return
	}

	r.sendResponse(writer, request, http.StatusOK, buf)
}

func (r *Router) createToken(writer http.ResponseWriter, request *http.Request) {
	name := request.URL.Query().Get("name")
	if name == "" {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, errEmptyName))

		return
	}

	token, err := r.tokenService.GenerateToken(request.Context(), name, false)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("CreateToken: %w", err))

		return
	}

	r.sendResponse(writer, request, http.StatusOK, []byte(token))
}

func (r *Router) createTeacherToken(writer http.ResponseWriter, request *http.Request) {
	name := request.URL.Query().Get("name")
	if name == "" {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrBadRequest, errEmptyName))

		return
	}

	token, err := r.tokenService.GenerateToken(request.Context(), name, true)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("CreateToken: %w", err))

		return
	}

	responseBody := TokenResponse{
		Token: token,
	}

	buf, err := json.Marshal(responseBody)
	if err != nil {
		r.sendErrorResponse(writer, request, fmt.Errorf("%w: %w", models.ErrInternalServer, err))

		return
	}

	r.sendResponse(writer, request, http.StatusOK, buf)
}

func getPaginationParameter(request *http.Request, parameterName string, defaultValue int) (int, error) {
	parameter := request.URL.Query().Get(parameterName)

	if parameter == "" {
		return defaultValue, nil
	}

	value, err := strconv.Atoi(parameter)
	if err != nil {
		return 0, fmt.Errorf("%w %s: %w", errInvalidPaginationParameter, parameterName, err)
	}

	if value <= 0 {
		return 0, fmt.Errorf("%w %s: %d", errInvalidPaginationParameter, parameterName, value)
	}

	return value, nil
}